refactor: move out reader and runner from service
